+++ sktime/datasets/_readers_writers/tsv.py
@@ -1,0 +1,43 @@
+"""Util function for reading tsv files."""
+
+__author__ = ["TonyBagnall"]
+__all__ = ["load_from_ucr_tsv_to_dataframe"]
+
+import pandas as pd
+
+
+# TODO: original author didnt add test for this function
+def load_from_ucr_tsv_to_dataframe(
+    full_file_path_and_name, return_separate_X_and_y=True
+):
+    """Load data from a .tsv file into a Pandas DataFrame.
+
+    Parameters
+    ----------
+    full_file_path_and_name: str
+        The full pathname of the .tsv file to read.
+    return_separate_X_and_y: bool
+        true then X and Y values should be returned as separate Data Frames (
+        X) and a numpy array (y), false otherwise.
+        This is only relevant for data.
+
+    Returns
+    -------
+    DataFrame, ndarray
+        If return_separate_X_and_y then a tuple containing a DataFrame and a
+        numpy array containing the relevant time-series and corresponding
+        class values.
+    DataFrame
+        If not return_separate_X_and_y then a single DataFrame containing
+        all time-series and (if relevant) a column "class_vals" the
+        associated class values.
+    """
+    df = pd.read_csv(full_file_path_and_name, sep="\t", header=None)
+    y = df.pop(0).values
+    df.columns -= 1
+    X = pd.DataFrame()
+    X["dim_0"] = [pd.Series(df.iloc[x, :]) for x in range(len(df))]
+    if return_separate_X_and_y is True:
+        return X, y
+    X["class_val"] = y
+    return X
